+++ subscriptions_service/src/main/java/com/subOne/subscriptions_service/kafka/kasfka_handler/KafkaHandlerService.java
@@ -1,5 +1,5 @@
 package com.subOne.subscriptions_service.kafka.kasfka_handler;
 
 public interface KafkaHandlerService {
-    void deleteSubscriptionsByGroup(Long groupId);
+    void deleteSubscriptionsByGroup(Integer groupId);
 }
