+++ subscriptions_service/src/main/java/com/subOne/subscriptions_service/service/AnalyticSubscriptionService.java
@@ -9,9 +9,9 @@ import org.springframework.security.oauth2.jwt.Jwt;
 import reactor.core.publisher.Mono;
 
 public interface AnalyticSubscriptionService {
-    Mono<ResponseTotalAnalyticSubscriptionDto> getTotalAnalyticById(Long groupId, Long subscriptionId, Jwt jwt);
-    Mono<ResponseAnalyticPaymentSubscriptionsDto> getPaymentInfoSubscriptionById(Long groupId, Long subscriptionId, Integer page, Jwt jwt);
-    Mono<ResponseTotalAnalyticSubscriptionGroupDto> getAlreadyPaidByGroupId(Long groupId, Jwt jwt);
+    Mono<ResponseTotalAnalyticSubscriptionDto> getTotalAnalyticById(Integer groupId, Integer subscriptionId, Jwt jwt);
+    Mono<ResponseAnalyticPaymentSubscriptionsDto> getPaymentInfoSubscriptionById(Integer groupId, Integer subscriptionId, Integer page, Jwt jwt);
+    Mono<ResponseTotalAnalyticSubscriptionGroupDto> getAlreadyPaidByGroupId(Integer groupId, Jwt jwt);
     Mono<ResponseTotalAnalyticGroupsDto> getTotalAnalyticGroups(Jwt jwt);
     Mono<Void> generateAnalyticSubscription(UserSubscription userSubscription);
 }
